+++ rasa_nlu/persistor.py
@@ -136,11 +136,9 @@ class AWSPersistor(Persistor):
 
     def __init__(self, aws_region, bucket_name, endpoint_url):
         # type: (Text, Text, Text) -> None
-
         super(AWSPersistor, self).__init__()
         self.s3 = boto3.resource('s3',
                                  region_name=aws_region,
-
                                  endpoint_url=endpoint_url)
         self._ensure_bucket_exists(bucket_name, aws_region)
         self.bucket_name = bucket_name