+++ CHANGELOG.rst
@@ -16,6 +16,8 @@ Added
 - support for splitting and training data into multiple and mixing formats
 - support for markdown files containing regex-features or synonyms only
 - added ability to list projects in cloud storage services for model loading
+- server evaluation endpoint at ``POST /evaluate``
+
 
 Changed
 -------